# swagger.yaml
openapi: 3.0.3
info:
  title: Retail POS Microservices API
  description: API documentation for the Retail Point of Sale (POS) backend microservices with AI-observable telemetry, multi-lane context headers, and chaos injection endpoints.
  version: 1.2.0
servers:
  - url: http://localhost:8081
    description: Order Service
  - url: http://localhost:8083
    description: Payment Service
  - url: http://localhost:8084
    description: Finalize Service
  - url: http://localhost:8085
    description: Telemetry Service
...
paths:
...
  /simulation/failures:
    post:
      summary: Chaos & Failure Injection Endpoint
      description: Inject synthetic timeouts or database failures for Dev Mode AI Ops testing (Available on ports 8081, 8083, 8084).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - type:
                - enabled:
              properties:
                type:
                  type: string
                  example: PAYMENT_TIMEOUT
                enabled:
                  type: boolean
                  example: true
                duration_ms:
                  type: integer
                  example: 5000
      responses:
        '200':
          description: Simulation failure state updated successfully
...
  /finalize:
    post:
      tags:
        - Finalize Service
      summary: Finalize transaction
      description: Verifies payment, checks stock availability, updates inventory, and marks order as COMPLETED.
      servers:
        - url: http://localhost:8084
      parameters:
        - $ref: '#/components/parameters/HeaderStoreID'
        - $ref: '#/components/parameters/HeaderLaneID'
        - $ref: '#/components/parameters/HeaderLaneType'
        - $ref: '#/components/parameters/HeaderCashierID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - order_no:
              properties:
                order_no:
                  type: string
                  example: ORD-10025
                trace_id:
                  type: string
                  example: TRC-82931
      responses:
        '200':
          description: Order finalized successfully
        '400':
          description: Order not paid or insufficient stock
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/POSError'
  /orders/{orderNo}/receipt:
    get:
      tags:
        - Finalize Service
      summary: Generate Receipt
      description: Generates a formatted text receipt for a COMPLETED order.
      servers:
        - url: http://localhost:8084
      parameters:
        - name: orderNo
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/HeaderStoreID'
        - $ref: '#/components/parameters/HeaderLaneID'
      responses:
        '200':
          description: Receipt generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  order_no:
                    type: string
                  receipt_text:
                    type: string
...
components:
  parameters:
    HeaderStoreID:
      name: X-Store-ID
      in: header
      required: false
      description: Store identifier (e.g. STORE-104)
      schema:
        type: string
        default: STORE-104
    HeaderLaneID:
      name: X-Lane-ID
      in: header
      required: false
      description: Checkout register lane ID (e.g. LANE-01, LANE-03)
      schema:
        type: string
        default: LANE-01
    HeaderLaneType:
      name: X-Lane-Type
      in: header
      required: false
      description: Checkout register lane type (CASHIER_EXPRESS, CASHIER_BELT, SELF_CHECKOUT)
      schema:
        type: string
        default: CASHIER_EXPRESS
    HeaderCashierID:
      name: X-Cashier-ID
      in: header
      required: false
      description: Operator / Cashier ID (e.g. CASHIER-101, CUSTOMER_SELF)
      schema:
        type: string
        default: CASHIER-101
  schemas:
...
    POSError:
      type: object
      properties:
        code:
          type: string
          example: "ERR_INSUFFICIENT_CASH"
        message:
          type: string
          example: "Insufficient cash received: Cash ($5.00) is less than Amount Due ($9.73)"
        category:
          type: string
          example: "BUSINESS"
...